@@ -1,49 +1,59 @@
 package com.l24o.workcalendar.modules
 
 import com.l24o.workcalendar.data.rest.models.Calendar
+import com.l24o.workcalendar.data.rest.models.Day
 import com.l24o.workcalendar.data.rest.models.Holiday
 import com.l24o.workcalendar.data.rest.models.TypeOfDay
 import org.threeten.bp.DayOfWeek
 import org.threeten.bp.LocalDate
-import org.threeten.bp.Month
-import org.threeten.bp.Year
+import java.io.Serializable
 import java.util.*
 
 object GodObject {
 
-    val mapOfHolidays: HashMap<Int, MutableList<Holiday>> = HashMap()
-    val mapOfHolidates: HashMap<Int, HashSet<LocalDate>> = HashMap()
-    val mapOfShortDates: HashMap<Int, HashSet<LocalDate>> = HashMap()
-
-    lateinit var calendar: Calendar
+    val mapOfHolidays: HashMap<YearWithMonth, MutableList<Holiday>> = HashMap()
+    val mapOfHolidates: HashMap<YearWithMonth, HashSet<LocalDate>> = HashMap()
+    val mapOfShortDates: HashMap<YearWithMonth, HashSet<LocalDate>> = HashMap()
+    val days = mutableListOf<Day>()
+    var minYear: Int = LocalDate.now().year
+    var maxYear: Int = LocalDate.now().year
 
     fun set(calendar: Calendar) {
-        this.calendar = calendar
-        fillData()
+        val year = calendar.year!!.toInt()
+        if (year > maxYear) {
+            maxYear = year
+        }
+        if (year < minYear) {
+            minYear = year
+        }
+        days.addAll(calendar.days)
+        fillData(calendar, year)
     }
 
-    private fun fillData() {
+    private fun fillData(calendar: Calendar, year: Int) {
         calendar.days.forEach { day ->
+            val yearWithMonth = YearWithMonth(
+                    year = year,
+                    month = day.day.monthValue
+            )
             day.holidayId?.takeIf { day.type == TypeOfDay.REST_DAY }?.let { holidayId ->
-                val monthNumber = day.day.monthValue
                 val holiday = calendar.holidays.first { it.id == holidayId }
                 holiday.date.add(day.day)
-                if (mapOfHolidays.containsKey(monthNumber)) {
-                    mapOfHolidays[monthNumber]?.let { list ->
+                if (mapOfHolidays.containsKey(yearWithMonth)) {
+                    mapOfHolidays[yearWithMonth]?.let { list ->
                         if (!list.contains(holiday)) {
-                            mapOfHolidays[monthNumber]?.add(holiday)
+                            mapOfHolidays[yearWithMonth]?.add(holiday)
                         }
                     }
                 } else {
                     val list = ArrayList<Holiday>()
                     list.add(holiday)
-                    mapOfHolidays.put(monthNumber, list)
+                    mapOfHolidays.put(yearWithMonth, list)
                 }
             }
             if (day.type == TypeOfDay.REST_DAY) {
-                val monthNumber = day.day.monthValue
-                if (mapOfHolidates.containsKey(monthNumber)) {
-                    mapOfHolidates[monthNumber]?.let { set ->
+                if (mapOfHolidates.containsKey(yearWithMonth)) {
+                    mapOfHolidates[yearWithMonth]?.let { set ->
                         if (!set.contains(day.day)) {
                             set.add(day.day)
                         }
@@ -51,13 +61,12 @@ object GodObject {
                 } else {
                     val set = HashSet<LocalDate>()
                     set.add(day.day)
-                    mapOfHolidates[monthNumber] = set
+                    mapOfHolidates[yearWithMonth] = set
                 }
             }
             if (day.type == TypeOfDay.SHORT_DAY) {
-                val monthNumber = day.day.monthValue
-                if (mapOfShortDates.containsKey(monthNumber)) {
-                    mapOfShortDates[monthNumber]?.let { set ->
+                if (mapOfShortDates.containsKey(yearWithMonth)) {
+                    mapOfShortDates[yearWithMonth]?.let { set ->
                         if (!set.contains(day.day)) {
                             set.add(day.day)
                         }
@@ -65,18 +74,22 @@ object GodObject {
                 } else {
                     val set = HashSet<LocalDate>()
                     set.add(day.day)
-                    mapOfShortDates[monthNumber] = set
+                    mapOfShortDates[yearWithMonth] = set
                 }
             }
         }
-        for (dayOfYear in 1..LocalDate.now().lengthOfYear()) {
-            val time = LocalDate.now().withDayOfYear(dayOfYear).atStartOfDay().toLocalDate()
+        val withYear = LocalDate.now().withYear(year)
+        for (dayOfYear in 1..withYear.lengthOfYear()) {
+            val time = withYear.withDayOfYear(dayOfYear).atStartOfDay().toLocalDate()
+            val yearWithMonth = YearWithMonth(
+                    year = year,
+                    month = time.monthValue
+            )
             val dayOfWeek = time.dayOfWeek
             if (dayOfWeek == DayOfWeek.SUNDAY || dayOfWeek == DayOfWeek.SATURDAY) {
-                val monthNumber = time.monthValue
-                if (mapOfShortDates[monthNumber] == null || mapOfShortDates[monthNumber]?.contains(time) == false) {
-                    if (mapOfHolidates.containsKey(monthNumber)) {
-                        mapOfHolidates[monthNumber]?.let { set ->
+                if (mapOfShortDates[yearWithMonth] == null || mapOfShortDates[yearWithMonth]?.contains(time) == false) {
+                    if (mapOfHolidates.containsKey(yearWithMonth)) {
+                        mapOfHolidates[yearWithMonth]?.let { set ->
                             if (!set.contains(time)) {
                                 set.add(time)
                             }
@@ -84,7 +97,7 @@ object GodObject {
                     } else {
                         val set = HashSet<LocalDate>()
                         set.add(time)
-                        mapOfHolidates[monthNumber] = set
+                        mapOfHolidates[yearWithMonth] = set
                     }
                 }
             }
@@ -93,6 +106,6 @@ object GodObject {
 }
 
 data class YearWithMonth(
-        val year: Year,
-        val month: Month
-)
+        var year: Int,
+        var month: Int
+) : Serializable
