@@ -3,13 +3,16 @@ package com.l24o.workcalendar.extensions
 import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.StaggeredGridLayoutManager
 
-
-fun RecyclerView.addOffsets(
+fun RecyclerView.addYearOffsets(
         offset: Int,
         columns: Int = 1
 ) {
     addItemDecoration(object : RecyclerView.ItemDecoration() {
+
+        var fullSizeSpan = hashSetOf<Int>()
+
         override fun getItemOffsets(
                 outRect: Rect,
                 view: View,
@@ -17,18 +20,32 @@ fun RecyclerView.addOffsets(
                 state: RecyclerView.State
         ) {
             val itemPosition = parent.getChildAdapterPosition(view)
-            val itemRow = itemPosition / columns
+            val isCurrentFullSizeSpan = (parent.getChildAt(itemPosition)?.layoutParams as? StaggeredGridLayoutManager.LayoutParams)?.isFullSpan
+                    ?: false
+            val isPrevFullSizeSpan = (parent.getChildAt(itemPosition - 1)?.layoutParams as? StaggeredGridLayoutManager.LayoutParams)?.isFullSpan
+                    ?: false
+            when {
+                isCurrentFullSizeSpan -> {
+                    fullSizeSpan.add(itemPosition)
+                    outRect.set(offset, offset, offset, offset)
+                }
+                !isPrevFullSizeSpan -> {
+                    val itemRow = (itemPosition - fullSizeSpan.size) / columns
 
-            val firstRow = itemRow == 0
-            val firstInRow = itemPosition % columns == 0
-            val lastInRow = itemPosition % columns == columns - 1
+                    val firstRow = itemRow == 0
+                    val firstInRow = (itemPosition - fullSizeSpan.size) % columns == 0
 
-            val left = if (firstInRow) offset else 0
-            val top = if (firstRow) offset else 0
-            val right = offset
-            val bottom = offset
+                    val left = if (firstInRow) offset else 0
+                    val top = if (firstRow) offset else 0
+                    val right = offset
+                    val bottom = offset
 
-            outRect.set(left, top, right, bottom)
+                    outRect.set(left, top, right, bottom)
+                }
+                isPrevFullSizeSpan -> {
+                    outRect.set(offset, 0, offset, offset)
+                }
+            }
         }
     })
 }
