@@ -9,12 +9,10 @@ import com.l24o.workcalendar.data.rest.models.Calendar
 import com.l24o.workcalendar.modules.GodObject
 import com.l24o.workcalendar.modules.calendar.CalendarActivity
 import com.redmadrobot.core.extension.schedulersIoToMain
-import io.reactivex.Completable
 import io.reactivex.Single
 import io.reactivex.disposables.Disposable
 import io.reactivex.functions.BiFunction
 import kotlinx.android.synthetic.main.activity_splash.*
-import org.threeten.bp.LocalDate
 import java.util.*
 import java.util.concurrent.TimeUnit
 
@@ -34,15 +32,10 @@ class SplashActivity : AppCompatActivity() {
         activity_splash_container_lottie.playAnimation()
 
         disposable = Single.zip(
-                CalendarAssetsManager.calendar(LocalDate.now().year),
+                CalendarAssetsManager.calendar().doOnSuccess { it.forEach { GodObject.set(it) } },
                 Single.timer(4, TimeUnit.SECONDS),
-                BiFunction<Calendar, Long, Calendar> { t1, _ -> t1 }
+                BiFunction<List<Calendar>, Long, List<Calendar>> { t1, _ -> t1 }
         )
-                .flatMapCompletable {
-                    Completable.fromCallable {
-                        GodObject.set(it)
-                    }
-                }
                 .schedulersIoToMain()
                 .subscribe({
                     startActivity(Intent(this@SplashActivity, CalendarActivity::class.java))
@@ -62,7 +55,7 @@ class SplashActivity : AppCompatActivity() {
     enum class ANIM(val resId: String) {
         ANIM_1("anim/anim_1.json"),
         ANIM_2("anim/anim_2.json"),
-//        ANIM_3("anim/anim_3.json"),
+        //        ANIM_3("anim/anim_3.json"),
         ANIM_5("anim/anim_5.json")
     }
 }
