@@ -14,12 +14,12 @@ class YearItem(
         val currentDate: LocalDate,
         val maximumDate: LocalDate,
         val minimumDate: LocalDate,
-        val onClickListener: () -> Unit
+        val onClickListener: (LocalDate) -> Unit
 ) : Item<ViewHolder>() {
     override fun getLayout() = R.layout.item_year
 
     override fun bind(viewHolder: ViewHolder, position: Int) {
-        viewHolder.itemView.item_year_container_click.setOnClickListener { onClickListener.invoke() }
+        viewHolder.itemView.item_year_container_click.setOnClickListener { onClickListener.invoke(currentDate) }
         viewHolder.itemView.item_year_text_view_title.text = viewHolder.itemView.context.resources.getStringArray(R.array.months).getOrNull(currentDate.monthValue - 1)
         with(viewHolder.itemView.item_year_calendar) {
             currentDate = CalendarDay.from(this@YearItem.currentDate)
@@ -34,11 +34,11 @@ class YearItem(
             addDecorators(
                     CurrentDayDecorator(context),
                     HighlightWeekendsDecorator(context),
-                    HoliDayDecorator(context, GodObject.calendar.days),
-                    ShortDayDecorator(context, GodObject.calendar.days),
+                    HoliDayDecorator(context, GodObject.days),
+                    ShortDayDecorator(context, GodObject.days),
                     HighlightWeekendsDecoratorWithToday(context),
-                    HoliDayDecoratorWithToday(context, GodObject.calendar.days),
-                    ShortDayDecoratorWithToday(context, GodObject.calendar.days)
+                    HoliDayDecoratorWithToday(context, GodObject.days),
+                    ShortDayDecoratorWithToday(context, GodObject.days)
             )
             selectionMode = MaterialCalendarView.SELECTION_MODE_NONE
         }
